[CMake] optionalize python
This package depends on Python to install 2 files. All the other files
can be installed without Python.

And we now have other packages (eg. hpp-pinocchio) which doesn't depend
on python, but which does depend on hpp-environments. So we need a way
to provide them what they need, without adding python as a required
dependency, while it is easily avoidable.

For this, with robotpkg, it is easy to build twice the same package, one
without anything dependent on python, and another one with only the
python stuff. This also allow us to install side-by-side the non-python
files, the python files for python 2, and the python files for python 3.

This commit allow this through 2 CMake options, while keeping everything
as it was by default.